% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{test}
\alias{test}
\title{Pipeline fun}
\usage{
test(path)
}
\arguments{
\item{path}{path to excel file}
}
\value{
dataframe with added `code` variable.
}
\description{
Reads excel table from file and uses the lookup function
to recode the values in `V1`
}
